fix locating now playing song in browser (using 'o')
due to change of keybinding for switching to local browser,
the trigger is invoked in SwitchTo() if main screen is already
pointing to browser. since we use 'o' in browser to locate now
playing song, this function was switching browser to wrong mode
and thus requested song couldn't be found.