status: reduce wakeups with display_bitrate enabled
When system was profiled for wakeups, a high wakeup rate from mpd was observed even when paused. Strace-ing mpd pointed at excessive polling even though mpd_communication_mode is set to notifications. By method of elimination in the configuration, it was narrowed down to display_bitrate. By code inspection, it could be seen that Mpd.isPlaying() is true even if mpd is paused, but it doesn't make sense as to why bitrate needs to be constantly updated even when mpd is paused. Hence the patch.

Signed-off-by: Raghavendra D Prabhu <rprabhu@wnohang.net>