Support locate of unknown tags/albums in media library
This adds support for locating a song in the media library, when the tag
or album (for respectively three and two column mode) is unknown. I.e.
the tag or album was not listed in the list of all tags or albums.

This is desirable when using a MPD server that is not able to give a
complete list of tags or albums. This may be the case for e.g. streaming
servers which does not have a list of the complete library.