Change active column if album/songs is empty
When you locate a song in the media library, if no albums or no songs
are found, set the active column to the previous one (tags if no albums,
and albums if no songs). This makes sure that the active column is not
empty, which may make it impossible to move out of.

The problem was if you highlight some song in the rightmost column in
the media browser and then go to some other window and select locate
song. If the tag or album it looked up in the media library was
empty, the selection would stay in the songs column while it was empty.
This made the selection impossible to change.

This only is a problem if a song has some tag or album for which the
find command doesn't return any results. This should never really happen
unless there is some inconsistency in the player. However, it may
happen, so we need to handle it.

Note: We don't want to return when no albums are found in two column
mode. In this case, we try to insert the album, as we do with tags when
they are not found.