song: Remove duplicates when combining tags
Some automatic taggers (especially beets) set multiple tags, that are
semantically identical, to the same value in order to maximize
compatibility. This results in ncmpcpp displaying rather strange looking
entries, like e.g. the track number "01 | 01".

This patch fixes that by removing duplicates when combining all tags to
a single string. Thus, the previous behavior is retained for tags that
are actually different.