Fix crash when jumping to playing song in playlist editor
By default, PlaylistEditor delays fetching of the content for performance
reasons. If the user switches the playlist then tells ncmpcpp to jump to the
currently playing song within the delay, the update request that fetches the new
content will be ignored. This results in inconsistent content, or crash if the
position of the song in the new content exceeds the length of the stale content.

This fix enforces a fetch even if the update request that fetches the new
content stays within the delay.