Add action to add streams via youtube-dl
youtube-dl [0] is a "command-line program to download videos from
YouTube.com" and many other websites, e.g. media libraries of public
broadcasters.

This commit adds a new action to ncmpcpp's playlist view: Pressing 'D',
prompts for the URL of a website.  The input is passed to youtube-dl to
retrieve the URL of the audio stream and available metadata.  If
successful, the URL is added to the MPD playlist.

[0]: https://github.com/ytdl-org/youtube-dl