Append newline to --current-song output
Currently the default behavior is to print information about the current
song without a trailing newline. This can make use in scripts rather
tedious as you need to manually append a newline to the output. This PR
appends a newline to the output which makes life easier for the user.